pyproject.toml
[tool.poetry]
name = "edward"
version = "0.1.0"
description = ""
authors = ["Rafael do Carmo <rafaelcarmo143@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
edward-bot = "edward.bot:start"
read-sheets = "edward.read_cheatsheets:run"
clear-meilisearch = "edward.read_cheatsheets:run_clear"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.2.2"
python-slugify = "^8.0.1"
python-dotenv = "^1.0.0"
openai = "^0.27.8"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
black = "^23.1.0"
pytest-cov = "^4.0.0"
isort = "^5.12.0"
taskipy = "^1.10.3"
mypy = "^1.3.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.1.3"
mkdocstrings = "^0.20.0"
mkdocstrings-python = "^0.8.3"
markdown-include = "^0.8.1"


[tool.poetry.group.infra.dependencies]
meilisearch = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "black --check --diff . && isort --check --diff ."
format = "black . && isort ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=edward -vv"
post_test = "coverage html"
start_bot = "poetry run edward-bot"